Move IsGrbViewContext to separate file
